requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.3
scikit-learn==1.3.2